server/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
PyJWT==2.9.0
